# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Offhours.ai API — AI-powered lifestyle productivity for busy professionals
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  clerkId: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  interests?: string[];
  onboardingComplete: boolean;
  /** @nullable */
  workAddress?: string | null;
  /** @nullable */
  workLat?: number | null;
  /** @nullable */
  workLng?: number | null;
  createdAt?: string;
}

export interface InterestsInput {
  /**
     * @minItems 5
     * @maxItems 15
     */
  interests: string[];
}

export interface LocationInput {
  /** Full street address of work location */
  workAddress: string;
}

export interface LocationResult {
  /** @nullable */
  workAddress?: string | null;
  /** @nullable */
  workLat?: number | null;
  /** @nullable */
  workLng?: number | null;
  /** Whether the address was successfully geocoded */
  geocoded?: boolean;
}

export interface CalendarGap {
  id: string;
  /** YYYY-MM-DD */
  date: string;
  /** Monday, Tuesday, etc. */
  dayOfWeek: string;
  /** HH:MM in 24h format e.g. 18:00 */
  startTime: string;
  /** HH:MM in 24h format e.g. 20:30 */
  endTime: string;
  durationMinutes: number;
  /**
     * Human-friendly label e.g. '2.5 hr window'
     * @nullable
     */
  label?: string | null;
}

export interface SuggestedEvent {
  id: string;
  title: string;
  /** e.g. Fitness, Food & Drink, Networking, Wellness */
  category: string;
  dayOfWeek: string;
  /** e.g. 'Tuesday 7:00 PM' */
  time: string;
  venue: string;
  neighborhood: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  matchedInterest?: string | null;
  /**
     * Estimated travel time in minutes from work address
     * @nullable
     */
  travelTimeMinutes?: number | null;
  /**
     * Human-readable travel label e.g. '12 min drive'
     * @nullable
     */
  travelLabel?: string | null;
  /**
     * Whether the event is reachable within the available gap
     * @nullable
     */
  travelable?: boolean | null;
}

export interface DashboardSummary {
  gapsThisWeek: number;
  totalFreeHours: number;
  suggestedEventsCount: number;
  /** YYYY-MM-DD of the Monday of current week */
  weekStartDate: string;
}