# lib/api-zod/src/generated/types/calendarGap.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Offhours.ai API — AI-powered lifestyle productivity for busy professionals
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarGap {
  id: string;
  /** YYYY-MM-DD */
  date: string;
  /** Monday, Tuesday, etc. */
  dayOfWeek: string;
  /** HH:MM in 24h format e.g. 18:00 */
  startTime: string;
  /** HH:MM in 24h format e.g. 20:30 */
  endTime: string;
  durationMinutes: number;
  /**
     * Human-friendly label e.g. '2.5 hr window'
     * @nullable
     */
  label?: string | null;
}