# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Offhours.ai API — AI-powered lifestyle productivity for busy professionals
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  gapsThisWeek: number;
  totalFreeHours: number;
  suggestedEventsCount: number;
  /** YYYY-MM-DD of the Monday of current week */
  weekStartDate: string;
}