# lib/api-zod/src/generated/types/locationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Offhours.ai API — AI-powered lifestyle productivity for busy professionals
 * OpenAPI spec version: 0.1.0
 */

export interface LocationInput {
  /** Full street address of work location */
  workAddress: string;
}