# lib/api-zod/src/generated/types/locationResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Offhours.ai API — AI-powered lifestyle productivity for busy professionals
 * OpenAPI spec version: 0.1.0
 */

export interface LocationResult {
  /** @nullable */
  workAddress?: string | null;
  /** @nullable */
  workLat?: number | null;
  /** @nullable */
  workLng?: number | null;
  /** Whether the address was successfully geocoded */
  geocoded?: boolean;
}