# lib/api-zod/src/generated/types/suggestedEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Offhours.ai API — AI-powered lifestyle productivity for busy professionals
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestedEvent {
  id: string;
  title: string;
  /** e.g. Fitness, Food & Drink, Networking, Wellness */
  category: string;
  dayOfWeek: string;
  /** e.g. 'Tuesday 7:00 PM' */
  time: string;
  venue: string;
  neighborhood: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  matchedInterest?: string | null;
  /**
     * Estimated travel time in minutes from work address
     * @nullable
     */
  travelTimeMinutes?: number | null;
  /**
     * Human-readable travel label e.g. '12 min drive'
     * @nullable
     */
  travelLabel?: string | null;
  /**
     * Whether the event is reachable within the available gap
     * @nullable
     */
  travelable?: boolean | null;
}